ADD: AdminController + Service and CustomUserController + service, DTO Classes
AdminController - changeUsername endpoint that takes in two usernames the first one is the user and the second is the new username.
changePassword endpoint takes in a username and changes its password to the one sent in + encodes it.

CustomUserController - getUsers endpoint that only takes users name and roles and returns it into json format, it works because i made a custom repository method that only takes in username and roles.